Type Definition earlgrey_registers::aes::ctrl::R

type R = R<u32, CTRL>;

Reader of register CTRL

Methods

impl R

pub fn mode(&self) -> MODE_R

Bit 0 - Select encryption(0) or decryption(1) operating mode of AES unit.

pub fn key_len(&self) -> KEY_LEN_R

Bits 1:3 - 3-bit one-hot field to select AES key length: 128 bit (3'b001), 192 bit (3'b010) or 256 bit (3'b100). Invalid input values, i.e., value with multiple bits set, value 3'b000, and value 3'b010 in case 192-bit keys are not supported (because disabled at compile time) are mapped to 3'b001.

pub fn manual_start_trigger(&self) -> MANUAL_START_TRIGGER_R

Bit 4 - Controls whether the AES unit should automatically start to encrypt/decrypt when it receives new input data (0) or wait for separate trigger signal before starting (1) (see Trigger Register).

pub fn force_data_overwrite(&self) -> FORCE_DATA_OVERWRITE_R

Bit 5 - Control whether the AES unit is stalled during the last encryption/decryption cycle if the previous output data has not yet been read (0) or finishes the operation and overwrites the previous output data (1).