Module earlgrey_registers::aes::data_in::data_in0

Input Data Registers. Loaded into the internal State register upon starting encryption/decryption of the next block. After that, the processor can update the Input Data Register. The order in which the registers are written does not matter.

Structs

DATA_IN0_W
Write proxy for field data_in0

Type Definitions

W
Writer for register DATA_IN0