Type Definition earlgrey_registers::aes::status::R

type R = R<u32, STATUS>;

Reader of register STATUS

Methods

impl R

pub fn idle(&self) -> IDLE_R

Bit 0 - The AES unit is idle (0) or busy (1).

pub fn stall(&self) -> STALL_R

Bit 1 - The AES unit is not stalled (0) or stalled (1) because there is previous output data that must be read by the processor before the AES unit can overwrite this data.

pub fn output_valid(&self) -> OUTPUT_VALID_R

Bit 2 - The AES unit has no valid output (0) or has valid output data (1).

pub fn input_ready(&self) -> INPUT_READY_R

Bit 3 - The AES unit is ready (1) to receive new data input via the DATA_IN registers or the present values in the DATA_IN registers have not yet been loaded into the module (0).