Module earlgrey_registers::alert_handler

alert_handler

Modules

alert_cause
Alert Cause Register

alert_class
Class assignment of alerts.

alert_en
Enable register for alerts.

classa_accum_cnt
Current accumulation value for alert Class A. Software can clear this register with a write to !!CLASSA_CLR register unless !!CLASSA_CLREN is false.

classa_accum_thresh
Accumulation threshold value for alert Class A.

classa_clr
Clear for esclation protocol of Class A.

classa_clren
Clear enable for escalation protocol of Class A alerts.

classa_ctrl
Escalation control register for alert Class A. Can not be modified if !!REGEN is false.

classa_esc_cnt
Escalation counter in cycles for Class A.

classa_phase0_cyc
Duration of escalation phase 0 for Class A.

classa_phase1_cyc
Duration of escalation phase 1 for Class A.

classa_phase2_cyc
Duration of escalation phase 2 for Class A.

classa_phase3_cyc
Duration of escalation phase 3 for Class A.

classa_state
Current escalation state of Class A. See also !!CLASSA_ESC_CNT.

classa_timeout_cyc
Interrupt timeout in cycles.

classb_accum_cnt
Current accumulation value for alert Class B. Software can clear this register with a write to !!CLASSB_CLR register unless !!CLASSB_CLREN is false.

classb_accum_thresh
Accumulation threshold value for alert Class B.

classb_clr
Clear for esclation protocol of Class B.

classb_clren
Clear enable for escalation protocol of Class B alerts.

classb_ctrl
Escalation control register for alert Class B. Can not be modified if !!REGEN is false.

classb_esc_cnt
Escalation counter in cycles for Class B.

classb_phase0_cyc
Duration of escalation phase 0 for Class B.

classb_phase1_cyc
Duration of escalation phase 1 for Class B.

classb_phase2_cyc
Duration of escalation phase 2 for Class B.

classb_phase3_cyc
Duration of escalation phase 3 for Class B.

classb_state
Current escalation state of Class B. See also !!CLASSB_ESC_CNT.

classb_timeout_cyc
Interrupt timeout in cycles.

classc_accum_cnt
Current accumulation value for alert Class C. Software can clear this register with a write to !!CLASSC_CLR register unless !!CLASSC_CLREN is false.

classc_accum_thresh
Accumulation threshold value for alert Class C.

classc_clr
Clear for esclation protocol of Class C.

classc_clren
Clear enable for escalation protocol of Class C alerts.

classc_ctrl
Escalation control register for alert Class C. Can not be modified if !!REGEN is false.

classc_esc_cnt
Escalation counter in cycles for Class C.

classc_phase0_cyc
Duration of escalation phase 0 for Class C.

classc_phase1_cyc
Duration of escalation phase 1 for Class C.

classc_phase2_cyc
Duration of escalation phase 2 for Class C.

classc_phase3_cyc
Duration of escalation phase 3 for Class C.

classc_state
Current escalation state of Class C. See also !!CLASSC_ESC_CNT.

classc_timeout_cyc
Interrupt timeout in cycles.

classd_accum_cnt
Current accumulation value for alert Class D. Software can clear this register with a write to !!CLASSD_CLR register unless !!CLASSD_CLREN is false.

classd_accum_thresh
Accumulation threshold value for alert Class D.

classd_clr
Clear for esclation protocol of Class D.

classd_clren
Clear enable for escalation protocol of Class D alerts.

classd_ctrl
Escalation control register for alert Class D. Can not be modified if !!REGEN is false.

classd_esc_cnt
Escalation counter in cycles for Class D.

classd_phase0_cyc
Duration of escalation phase 0 for Class D.

classd_phase1_cyc
Duration of escalation phase 1 for Class D.

classd_phase2_cyc
Duration of escalation phase 2 for Class D.

classd_phase3_cyc
Duration of escalation phase 3 for Class D.

classd_state
Current escalation state of Class D. See also !!CLASSD_ESC_CNT.

classd_timeout_cyc
Interrupt timeout in cycles.

intr_enable
Interrupt Enable Register

intr_state
Interrupt State Register

intr_test
Interrupt Test Register

loc_alert_cause
Alert Cause Register for Local Alerts. "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

loc_alert_class
Class assignment of local alerts. "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

loc_alert_en
Enable register for the aggregated local alerts "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

ping_timeout_cyc
Ping timeout cycle count.

regen
Register write enable for all control registers.

Structs

RegisterBlock
Register block

Type Definitions

ALERT_CAUSE
Alert Cause Register

ALERT_CLASS
Class assignment of alerts.

ALERT_EN
Enable register for alerts.

CLASSA_ACCUM_CNT
Current accumulation value for alert Class A. Software can clear this register with a write to !!CLASSA_CLR register unless !!CLASSA_CLREN is false.

CLASSA_ACCUM_THRESH
Accumulation threshold value for alert Class A.

CLASSA_CLR
Clear for esclation protocol of Class A.

CLASSA_CLREN
Clear enable for escalation protocol of Class A alerts.

CLASSA_CTRL
Escalation control register for alert Class A. Can not be modified if !!REGEN is false.

CLASSA_ESC_CNT
Escalation counter in cycles for Class A.

CLASSA_PHASE0_CYC
Duration of escalation phase 0 for Class A.

CLASSA_PHASE1_CYC
Duration of escalation phase 1 for Class A.

CLASSA_PHASE2_CYC
Duration of escalation phase 2 for Class A.

CLASSA_PHASE3_CYC
Duration of escalation phase 3 for Class A.

CLASSA_STATE
Current escalation state of Class A. See also !!CLASSA_ESC_CNT.

CLASSA_TIMEOUT_CYC
Interrupt timeout in cycles.

CLASSB_ACCUM_CNT
Current accumulation value for alert Class B. Software can clear this register with a write to !!CLASSB_CLR register unless !!CLASSB_CLREN is false.

CLASSB_ACCUM_THRESH
Accumulation threshold value for alert Class B.

CLASSB_CLR
Clear for esclation protocol of Class B.

CLASSB_CLREN
Clear enable for escalation protocol of Class B alerts.

CLASSB_CTRL
Escalation control register for alert Class B. Can not be modified if !!REGEN is false.

CLASSB_ESC_CNT
Escalation counter in cycles for Class B.

CLASSB_PHASE0_CYC
Duration of escalation phase 0 for Class B.

CLASSB_PHASE1_CYC
Duration of escalation phase 1 for Class B.

CLASSB_PHASE2_CYC
Duration of escalation phase 2 for Class B.

CLASSB_PHASE3_CYC
Duration of escalation phase 3 for Class B.

CLASSB_STATE
Current escalation state of Class B. See also !!CLASSB_ESC_CNT.

CLASSB_TIMEOUT_CYC
Interrupt timeout in cycles.

CLASSC_ACCUM_CNT
Current accumulation value for alert Class C. Software can clear this register with a write to !!CLASSC_CLR register unless !!CLASSC_CLREN is false.

CLASSC_ACCUM_THRESH
Accumulation threshold value for alert Class C.

CLASSC_CLR
Clear for esclation protocol of Class C.

CLASSC_CLREN
Clear enable for escalation protocol of Class C alerts.

CLASSC_CTRL
Escalation control register for alert Class C. Can not be modified if !!REGEN is false.

CLASSC_ESC_CNT
Escalation counter in cycles for Class C.

CLASSC_PHASE0_CYC
Duration of escalation phase 0 for Class C.

CLASSC_PHASE1_CYC
Duration of escalation phase 1 for Class C.

CLASSC_PHASE2_CYC
Duration of escalation phase 2 for Class C.

CLASSC_PHASE3_CYC
Duration of escalation phase 3 for Class C.

CLASSC_STATE
Current escalation state of Class C. See also !!CLASSC_ESC_CNT.

CLASSC_TIMEOUT_CYC
Interrupt timeout in cycles.

CLASSD_ACCUM_CNT
Current accumulation value for alert Class D. Software can clear this register with a write to !!CLASSD_CLR register unless !!CLASSD_CLREN is false.

CLASSD_ACCUM_THRESH
Accumulation threshold value for alert Class D.

CLASSD_CLR
Clear for esclation protocol of Class D.

CLASSD_CLREN
Clear enable for escalation protocol of Class D alerts.

CLASSD_CTRL
Escalation control register for alert Class D. Can not be modified if !!REGEN is false.

CLASSD_ESC_CNT
Escalation counter in cycles for Class D.

CLASSD_PHASE0_CYC
Duration of escalation phase 0 for Class D.

CLASSD_PHASE1_CYC
Duration of escalation phase 1 for Class D.

CLASSD_PHASE2_CYC
Duration of escalation phase 2 for Class D.

CLASSD_PHASE3_CYC
Duration of escalation phase 3 for Class D.

CLASSD_STATE
Current escalation state of Class D. See also !!CLASSD_ESC_CNT.

CLASSD_TIMEOUT_CYC
Interrupt timeout in cycles.

INTR_ENABLE
Interrupt Enable Register

INTR_STATE
Interrupt State Register

INTR_TEST
Interrupt Test Register

LOC_ALERT_CAUSE
Alert Cause Register for Local Alerts. "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

LOC_ALERT_CLASS
Class assignment of local alerts. "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

LOC_ALERT_EN
Enable register for the aggregated local alerts "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

PING_TIMEOUT_CYC
Ping timeout cycle count.

REGEN
Register write enable for all control registers.