Type Definition earlgrey_registers::alert_handler::classc_ctrl::R

type R = R<u32, CLASSC_CTRL>;

Reader of register CLASSC_CTRL

Methods

impl R

pub fn en(&self) -> EN_R

Bit 0 - Enable escalation mechanisms (accumulation and interrupt timeout) for Class C. Note that interrupts can fire regardless of whether the escalation mechanisms are enabled for this class or not.

pub fn lock(&self) -> LOCK_R

Bit 1 - Enable automatic locking of escalation counter for class C. If true, there is no way to stop the escalation protocol for class C once it has been triggered.

pub fn en_e0(&self) -> EN_E0_R

Bit 2 - Enable escalation signal 0 for Class C

pub fn en_e1(&self) -> EN_E1_R

Bit 3 - Enable escalation signal 1 for Class C

pub fn en_e2(&self) -> EN_E2_R

Bit 4 - Enable escalation signal 2 for Class C

pub fn en_e3(&self) -> EN_E3_R

Bit 5 - Enable escalation signal 3 for Class C

pub fn map_e0(&self) -> MAP_E0_R

Bits 6:7 - Determines in which escalation phase escalation signal 0 shall be asserted.

pub fn map_e1(&self) -> MAP_E1_R

Bits 8:9 - Determines in which escalation phase escalation signal 1 shall be asserted.

pub fn map_e2(&self) -> MAP_E2_R

Bits 10:11 - Determine sin which escalation phase escalation signal 2 shall be asserted.

pub fn map_e3(&self) -> MAP_E3_R

Bits 12:13 - Determines in which escalation phase escalation signal 3 shall be asserted.