Struct earlgrey_registers::alert_handler::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {
    pub intr_state: INTR_STATE,
    pub intr_enable: INTR_ENABLE,
    pub intr_test: INTR_TEST,
    pub regen: REGEN,
    pub ping_timeout_cyc: PING_TIMEOUT_CYC,
    pub alert_en: ALERT_EN,
    pub alert_class: ALERT_CLASS,
    pub alert_cause: ALERT_CAUSE,
    pub loc_alert_en: LOC_ALERT_EN,
    pub loc_alert_class: LOC_ALERT_CLASS,
    pub loc_alert_cause: LOC_ALERT_CAUSE,
    pub classa_ctrl: CLASSA_CTRL,
    pub classa_clren: CLASSA_CLREN,
    pub classa_clr: CLASSA_CLR,
    pub classa_accum_cnt: CLASSA_ACCUM_CNT,
    pub classa_accum_thresh: CLASSA_ACCUM_THRESH,
    pub classa_timeout_cyc: CLASSA_TIMEOUT_CYC,
    pub classa_phase0_cyc: CLASSA_PHASE0_CYC,
    pub classa_phase1_cyc: CLASSA_PHASE1_CYC,
    pub classa_phase2_cyc: CLASSA_PHASE2_CYC,
    pub classa_phase3_cyc: CLASSA_PHASE3_CYC,
    pub classa_esc_cnt: CLASSA_ESC_CNT,
    pub classa_state: CLASSA_STATE,
    pub classb_ctrl: CLASSB_CTRL,
    pub classb_clren: CLASSB_CLREN,
    pub classb_clr: CLASSB_CLR,
    pub classb_accum_cnt: CLASSB_ACCUM_CNT,
    pub classb_accum_thresh: CLASSB_ACCUM_THRESH,
    pub classb_timeout_cyc: CLASSB_TIMEOUT_CYC,
    pub classb_phase0_cyc: CLASSB_PHASE0_CYC,
    pub classb_phase1_cyc: CLASSB_PHASE1_CYC,
    pub classb_phase2_cyc: CLASSB_PHASE2_CYC,
    pub classb_phase3_cyc: CLASSB_PHASE3_CYC,
    pub classb_esc_cnt: CLASSB_ESC_CNT,
    pub classb_state: CLASSB_STATE,
    pub classc_ctrl: CLASSC_CTRL,
    pub classc_clren: CLASSC_CLREN,
    pub classc_clr: CLASSC_CLR,
    pub classc_accum_cnt: CLASSC_ACCUM_CNT,
    pub classc_accum_thresh: CLASSC_ACCUM_THRESH,
    pub classc_timeout_cyc: CLASSC_TIMEOUT_CYC,
    pub classc_phase0_cyc: CLASSC_PHASE0_CYC,
    pub classc_phase1_cyc: CLASSC_PHASE1_CYC,
    pub classc_phase2_cyc: CLASSC_PHASE2_CYC,
    pub classc_phase3_cyc: CLASSC_PHASE3_CYC,
    pub classc_esc_cnt: CLASSC_ESC_CNT,
    pub classc_state: CLASSC_STATE,
    pub classd_ctrl: CLASSD_CTRL,
    pub classd_clren: CLASSD_CLREN,
    pub classd_clr: CLASSD_CLR,
    pub classd_accum_cnt: CLASSD_ACCUM_CNT,
    pub classd_accum_thresh: CLASSD_ACCUM_THRESH,
    pub classd_timeout_cyc: CLASSD_TIMEOUT_CYC,
    pub classd_phase0_cyc: CLASSD_PHASE0_CYC,
    pub classd_phase1_cyc: CLASSD_PHASE1_CYC,
    pub classd_phase2_cyc: CLASSD_PHASE2_CYC,
    pub classd_phase3_cyc: CLASSD_PHASE3_CYC,
    pub classd_esc_cnt: CLASSD_ESC_CNT,
    pub classd_state: CLASSD_STATE,
    // some fields omitted
}

Register block

Fields
intr_state: INTR_STATE

0x00 - Interrupt State Register

intr_enable: INTR_ENABLE

0x04 - Interrupt Enable Register

intr_test: INTR_TEST

0x08 - Interrupt Test Register

regen: REGEN

0x0c - Register write enable for all control registers.

ping_timeout_cyc: PING_TIMEOUT_CYC

0x10 - Ping timeout cycle count.

alert_en: ALERT_EN

0x14 - Enable register for alerts.

alert_class: ALERT_CLASS

0x18 - Class assignment of alerts.

alert_cause: ALERT_CAUSE

0x1c - Alert Cause Register

loc_alert_en: LOC_ALERT_EN

0x20 - Enable register for the aggregated local alerts "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

loc_alert_class: LOC_ALERT_CLASS

0x24 - Class assignment of local alerts. "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

loc_alert_cause: LOC_ALERT_CAUSE

0x28 - Alert Cause Register for Local Alerts. "alert pingfail" (0), "escalation pingfail" (1), "alert integfail" (2) and "escalation integfail" (3).

classa_ctrl: CLASSA_CTRL

0x2c - Escalation control register for alert Class A. Can not be modified if !!REGEN is false.

classa_clren: CLASSA_CLREN

0x30 - Clear enable for escalation protocol of Class A alerts.

classa_clr: CLASSA_CLR

0x34 - Clear for esclation protocol of Class A.

classa_accum_cnt: CLASSA_ACCUM_CNT

0x38 - Current accumulation value for alert Class A. Software can clear this register with a write to !!CLASSA_CLR register unless !!CLASSA_CLREN is false.

classa_accum_thresh: CLASSA_ACCUM_THRESH

0x3c - Accumulation threshold value for alert Class A.

classa_timeout_cyc: CLASSA_TIMEOUT_CYC

0x40 - Interrupt timeout in cycles.

classa_phase0_cyc: CLASSA_PHASE0_CYC

0x44 - Duration of escalation phase 0 for Class A.

classa_phase1_cyc: CLASSA_PHASE1_CYC

0x48 - Duration of escalation phase 1 for Class A.

classa_phase2_cyc: CLASSA_PHASE2_CYC

0x4c - Duration of escalation phase 2 for Class A.

classa_phase3_cyc: CLASSA_PHASE3_CYC

0x50 - Duration of escalation phase 3 for Class A.

classa_esc_cnt: CLASSA_ESC_CNT

0x54 - Escalation counter in cycles for Class A.

classa_state: CLASSA_STATE

0x58 - Current escalation state of Class A. See also !!CLASSA_ESC_CNT.

classb_ctrl: CLASSB_CTRL

0x5c - Escalation control register for alert Class B. Can not be modified if !!REGEN is false.

classb_clren: CLASSB_CLREN

0x60 - Clear enable for escalation protocol of Class B alerts.

classb_clr: CLASSB_CLR

0x64 - Clear for esclation protocol of Class B.

classb_accum_cnt: CLASSB_ACCUM_CNT

0x68 - Current accumulation value for alert Class B. Software can clear this register with a write to !!CLASSB_CLR register unless !!CLASSB_CLREN is false.

classb_accum_thresh: CLASSB_ACCUM_THRESH

0x6c - Accumulation threshold value for alert Class B.

classb_timeout_cyc: CLASSB_TIMEOUT_CYC

0x70 - Interrupt timeout in cycles.

classb_phase0_cyc: CLASSB_PHASE0_CYC

0x74 - Duration of escalation phase 0 for Class B.

classb_phase1_cyc: CLASSB_PHASE1_CYC

0x78 - Duration of escalation phase 1 for Class B.

classb_phase2_cyc: CLASSB_PHASE2_CYC

0x7c - Duration of escalation phase 2 for Class B.

classb_phase3_cyc: CLASSB_PHASE3_CYC

0x80 - Duration of escalation phase 3 for Class B.

classb_esc_cnt: CLASSB_ESC_CNT

0x84 - Escalation counter in cycles for Class B.

classb_state: CLASSB_STATE

0x88 - Current escalation state of Class B. See also !!CLASSB_ESC_CNT.

classc_ctrl: CLASSC_CTRL

0x8c - Escalation control register for alert Class C. Can not be modified if !!REGEN is false.

classc_clren: CLASSC_CLREN

0x90 - Clear enable for escalation protocol of Class C alerts.

classc_clr: CLASSC_CLR

0x94 - Clear for esclation protocol of Class C.

classc_accum_cnt: CLASSC_ACCUM_CNT

0x98 - Current accumulation value for alert Class C. Software can clear this register with a write to !!CLASSC_CLR register unless !!CLASSC_CLREN is false.

classc_accum_thresh: CLASSC_ACCUM_THRESH

0x9c - Accumulation threshold value for alert Class C.

classc_timeout_cyc: CLASSC_TIMEOUT_CYC

0xa0 - Interrupt timeout in cycles.

classc_phase0_cyc: CLASSC_PHASE0_CYC

0xa4 - Duration of escalation phase 0 for Class C.

classc_phase1_cyc: CLASSC_PHASE1_CYC

0xa8 - Duration of escalation phase 1 for Class C.

classc_phase2_cyc: CLASSC_PHASE2_CYC

0xac - Duration of escalation phase 2 for Class C.

classc_phase3_cyc: CLASSC_PHASE3_CYC

0xb0 - Duration of escalation phase 3 for Class C.

classc_esc_cnt: CLASSC_ESC_CNT

0xb4 - Escalation counter in cycles for Class C.

classc_state: CLASSC_STATE

0xb8 - Current escalation state of Class C. See also !!CLASSC_ESC_CNT.

classd_ctrl: CLASSD_CTRL

0xbc - Escalation control register for alert Class D. Can not be modified if !!REGEN is false.

classd_clren: CLASSD_CLREN

0xc0 - Clear enable for escalation protocol of Class D alerts.

classd_clr: CLASSD_CLR

0xc4 - Clear for esclation protocol of Class D.

classd_accum_cnt: CLASSD_ACCUM_CNT

0xc8 - Current accumulation value for alert Class D. Software can clear this register with a write to !!CLASSD_CLR register unless !!CLASSD_CLREN is false.

classd_accum_thresh: CLASSD_ACCUM_THRESH

0xcc - Accumulation threshold value for alert Class D.

classd_timeout_cyc: CLASSD_TIMEOUT_CYC

0xd0 - Interrupt timeout in cycles.

classd_phase0_cyc: CLASSD_PHASE0_CYC

0xd4 - Duration of escalation phase 0 for Class D.

classd_phase1_cyc: CLASSD_PHASE1_CYC

0xd8 - Duration of escalation phase 1 for Class D.

classd_phase2_cyc: CLASSD_PHASE2_CYC

0xdc - Duration of escalation phase 2 for Class D.

classd_phase3_cyc: CLASSD_PHASE3_CYC

0xe0 - Duration of escalation phase 3 for Class D.

classd_esc_cnt: CLASSD_ESC_CNT

0xe4 - Escalation counter in cycles for Class D.

classd_state: CLASSD_STATE

0xe8 - Current escalation state of Class D. See also !!CLASSD_ESC_CNT.