Type Definition earlgrey_registers::flash_ctrl::control::R

type R = R<u32, CONTROL>;

Reader of register CONTROL

Methods

impl R

pub fn start(&self) -> START_R

Bit 0 - Start flash transaction. This bit shall only be set after the other fields of the CONTROL register and ADDR have been programmed

pub fn op(&self) -> OP_R

Bits 4:5 - Flash operation selection

pub fn erase_sel(&self) -> ERASE_SEL_R

Bit 6 - Flash operation selection

pub fn fifo_rst(&self) -> FIFO_RST_R

Bit 7 - RST FIFOs

pub fn num(&self) -> NUM_R

Bits 16:27 - Number of flash words the flash operation should read or program.