Module earlgrey_registers::hmac

hmac

Modules

cfg
HMAC Configuration register. The register is updated when the engine is in Idle. If the software updates the register while the engine computes the hash, the updated value is discarded.

cmd
HMAC command register

digest
Register block Digest output. If HMAC is disabled, the register shows result of SHA256 Order of the digest is: digest[255:0] = {DIGEST0, DIGEST1, DIGEST2, ... , DIGEST7};

err_code
HMAC Error Code

intr_enable
Interrupt Enable Register

intr_state
Interrupt State Register

intr_test
Interrupt Test Register

key
Register block HMAC Secret Key SHA256 assumes secret key is hashed 256bit key. Order of the secret key is: key[255:0] = {KEY0, KEY1, KEY2, ... , KEY7}; The registers are allowed to be updated when the engine is in Idle state. If the engine computes the hash, it discards any attempts to update the secret keys and report an error.

msg_fifo
Message FIFO. Any write to this window will be appended to the FIFO. Only the lower [1:0] bits of the address matter to writes within the window (for correctly dealing with non 32-bit writes)

msg_length_lower
Received Message Length calculated by the HMAC in bits [31:0] Message is byte granularity. lower 3bits [2:0] are ignored.

msg_length_upper
Received Message Length calculated by the HMAC in bits [63:32]

status
HMAC Status register

wipe_secret
Randomize internal secret registers. If CPU writes value into the register, the value is used to randomize internal variables such as secret key, internal state machine, or hash value.

Structs

DIGEST
Register block

KEY
Register block

RegisterBlock
Register block

Type Definitions

CFG
HMAC Configuration register. The register is updated when the engine is in Idle. If the software updates the register while the engine computes the hash, the updated value is discarded.

CMD
HMAC command register

ERR_CODE
HMAC Error Code

INTR_ENABLE
Interrupt Enable Register

INTR_STATE
Interrupt State Register

INTR_TEST
Interrupt Test Register

MSG_FIFO
Message FIFO. Any write to this window will be appended to the FIFO. Only the lower [1:0] bits of the address matter to writes within the window (for correctly dealing with non 32-bit writes)

MSG_LENGTH_LOWER
Received Message Length calculated by the HMAC in bits [31:0] Message is byte granularity. lower 3bits [2:0] are ignored.

MSG_LENGTH_UPPER
Received Message Length calculated by the HMAC in bits [63:32]

STATUS
HMAC Status register

WIPE_SECRET
Randomize internal secret registers. If CPU writes value into the register, the value is used to randomize internal variables such as secret key, internal state machine, or hash value.