Module earlgrey_registers::hmac::cfg

HMAC Configuration register. The register is updated when the engine is in Idle. If the software updates the register while the engine computes the hash, the updated value is discarded.

Structs

DIGEST_SWAP_W
Write proxy for field digest_swap

ENDIAN_SWAP_W
Write proxy for field endian_swap

HMAC_EN_W
Write proxy for field hmac_en

SHA_EN_W
Write proxy for field sha_en

Type Definitions

DIGEST_SWAP_R
Reader of field digest_swap

ENDIAN_SWAP_R
Reader of field endian_swap

HMAC_EN_R
Reader of field hmac_en

R
Reader of register CFG

SHA_EN_R
Reader of field sha_en

W
Writer for register CFG