Module earlgrey_registers::rv_plic

rv_plic

Modules

cc0
Claim interrupt by read, complete interrupt by write for Target 0. Value read/written is interrupt ID. Reading a value of 0 means no pending interrupts.

ie0
Register block Interrupt Enable for Target 0

ip
Register block Interrupt Pending

le
Register block Interrupt Source mode. 0: Level, 1: Edge-triggered

msip0
msip for Hart 0. Write 1 to here asserts software interrupt for Hart msip_o[0], write 0 to clear.

prio0
Interrupt Source 0 Priority

prio1
Interrupt Source 1 Priority

prio2
Interrupt Source 2 Priority

prio3
Interrupt Source 3 Priority

prio4
Interrupt Source 4 Priority

prio5
Interrupt Source 5 Priority

prio6
Interrupt Source 6 Priority

prio7
Interrupt Source 7 Priority

prio8
Interrupt Source 8 Priority

prio9
Interrupt Source 9 Priority

prio10
Interrupt Source 10 Priority

prio11
Interrupt Source 11 Priority

prio12
Interrupt Source 12 Priority

prio13
Interrupt Source 13 Priority

prio14
Interrupt Source 14 Priority

prio15
Interrupt Source 15 Priority

prio16
Interrupt Source 16 Priority

prio17
Interrupt Source 17 Priority

prio18
Interrupt Source 18 Priority

prio19
Interrupt Source 19 Priority

prio20
Interrupt Source 20 Priority

prio21
Interrupt Source 21 Priority

prio22
Interrupt Source 22 Priority

prio23
Interrupt Source 23 Priority

prio24
Interrupt Source 24 Priority

prio25
Interrupt Source 25 Priority

prio26
Interrupt Source 26 Priority

prio27
Interrupt Source 27 Priority

prio28
Interrupt Source 28 Priority

prio29
Interrupt Source 29 Priority

prio30
Interrupt Source 30 Priority

prio31
Interrupt Source 31 Priority

prio32
Interrupt Source 32 Priority

prio33
Interrupt Source 33 Priority

prio34
Interrupt Source 34 Priority

prio35
Interrupt Source 35 Priority

prio36
Interrupt Source 36 Priority

prio37
Interrupt Source 37 Priority

prio38
Interrupt Source 38 Priority

prio39
Interrupt Source 39 Priority

prio40
Interrupt Source 40 Priority

prio41
Interrupt Source 41 Priority

prio42
Interrupt Source 42 Priority

prio43
Interrupt Source 43 Priority

prio44
Interrupt Source 44 Priority

prio45
Interrupt Source 45 Priority

prio46
Interrupt Source 46 Priority

prio47
Interrupt Source 47 Priority

prio48
Interrupt Source 48 Priority

prio49
Interrupt Source 49 Priority

prio50
Interrupt Source 50 Priority

prio51
Interrupt Source 51 Priority

prio52
Interrupt Source 52 Priority

prio53
Interrupt Source 53 Priority

prio54
Interrupt Source 54 Priority

prio55
Interrupt Source 55 Priority

prio56
Interrupt Source 56 Priority

prio57
Interrupt Source 57 Priority

prio58
Interrupt Source 58 Priority

prio59
Interrupt Source 59 Priority

prio60
Interrupt Source 60 Priority

prio61
Interrupt Source 61 Priority

prio62
Interrupt Source 62 Priority

threshold0
Threshold of priority for Target 0

Structs

IE0
Register block

IP
Register block

LE
Register block

RegisterBlock
Register block

Type Definitions

CC0
Claim interrupt by read, complete interrupt by write for Target 0. Value read/written is interrupt ID. Reading a value of 0 means no pending interrupts.

MSIP0
msip for Hart 0. Write 1 to here asserts software interrupt for Hart msip_o[0], write 0 to clear.

PRIO0
Interrupt Source 0 Priority

PRIO1
Interrupt Source 1 Priority

PRIO2
Interrupt Source 2 Priority

PRIO3
Interrupt Source 3 Priority

PRIO4
Interrupt Source 4 Priority

PRIO5
Interrupt Source 5 Priority

PRIO6
Interrupt Source 6 Priority

PRIO7
Interrupt Source 7 Priority

PRIO8
Interrupt Source 8 Priority

PRIO9
Interrupt Source 9 Priority

PRIO10
Interrupt Source 10 Priority

PRIO11
Interrupt Source 11 Priority

PRIO12
Interrupt Source 12 Priority

PRIO13
Interrupt Source 13 Priority

PRIO14
Interrupt Source 14 Priority

PRIO15
Interrupt Source 15 Priority

PRIO16
Interrupt Source 16 Priority

PRIO17
Interrupt Source 17 Priority

PRIO18
Interrupt Source 18 Priority

PRIO19
Interrupt Source 19 Priority

PRIO20
Interrupt Source 20 Priority

PRIO21
Interrupt Source 21 Priority

PRIO22
Interrupt Source 22 Priority

PRIO23
Interrupt Source 23 Priority

PRIO24
Interrupt Source 24 Priority

PRIO25
Interrupt Source 25 Priority

PRIO26
Interrupt Source 26 Priority

PRIO27
Interrupt Source 27 Priority

PRIO28
Interrupt Source 28 Priority

PRIO29
Interrupt Source 29 Priority

PRIO30
Interrupt Source 30 Priority

PRIO31
Interrupt Source 31 Priority

PRIO32
Interrupt Source 32 Priority

PRIO33
Interrupt Source 33 Priority

PRIO34
Interrupt Source 34 Priority

PRIO35
Interrupt Source 35 Priority

PRIO36
Interrupt Source 36 Priority

PRIO37
Interrupt Source 37 Priority

PRIO38
Interrupt Source 38 Priority

PRIO39
Interrupt Source 39 Priority

PRIO40
Interrupt Source 40 Priority

PRIO41
Interrupt Source 41 Priority

PRIO42
Interrupt Source 42 Priority

PRIO43
Interrupt Source 43 Priority

PRIO44
Interrupt Source 44 Priority

PRIO45
Interrupt Source 45 Priority

PRIO46
Interrupt Source 46 Priority

PRIO47
Interrupt Source 47 Priority

PRIO48
Interrupt Source 48 Priority

PRIO49
Interrupt Source 49 Priority

PRIO50
Interrupt Source 50 Priority

PRIO51
Interrupt Source 51 Priority

PRIO52
Interrupt Source 52 Priority

PRIO53
Interrupt Source 53 Priority

PRIO54
Interrupt Source 54 Priority

PRIO55
Interrupt Source 55 Priority

PRIO56
Interrupt Source 56 Priority

PRIO57
Interrupt Source 57 Priority

PRIO58
Interrupt Source 58 Priority

PRIO59
Interrupt Source 59 Priority

PRIO60
Interrupt Source 60 Priority

PRIO61
Interrupt Source 61 Priority

PRIO62
Interrupt Source 62 Priority

THRESHOLD0
Threshold of priority for Target 0