Type Definition earlgrey_registers::spi_device::control::W

type W = W<u32, CONTROL>;

Writer for register CONTROL

Methods

impl W

pub fn abort(&mut self) -> ABORT_W

Bit 0 - Abort pending jobs. If TX_FIFO (async) is full, this command can let TXF Control logic back to Idle state

pub fn mode(&mut self) -> MODE_W

Bits 4:5 - SPI Device operation mode. Currently only FwMode is supported.

pub fn rst_txfifo(&mut self) -> RST_TXFIFO_W

Bit 16 - Reset Async TX_FIFO. This only resets asynchronous fifo. If firmware wants to reset SRAM FIFO, it should write 0 into read/write pointers. Note: This value should be controlled only when SPI interface is in Idle state as this reset signal doesn't have reset synchronizer.

pub fn rst_rxfifo(&mut self) -> RST_RXFIFO_W

Bit 17 - Reset Async RX_FIFO. This only resets asynchronous fifo. If firmware wants to reset SRAM FIFO, it should write 0 into read pointer and write pointer. Note: This value should be controlled only when SPI interface is in Idle state as this reset signal doesn't have reset synchronizer.