Type Definition earlgrey_registers::uart::ctrl::R

type R = R<u32, CTRL>;

Reader of register CTRL

Methods

impl R

pub fn tx(&self) -> TX_R

Bit 0 - TX enable

pub fn rx(&self) -> RX_R

Bit 1 - RX enable

pub fn nf(&self) -> NF_R

Bit 2 - RX noise filter enable. If the noise filter is enabled, RX line goes through the 3-tap repetition code. It ignores single IP clock period noise.

pub fn slpbk(&self) -> SLPBK_R

Bit 4 - System loopback enable. If this bit is turned on, any outgoing bits to TX are received through RX. See Block Diagram. Note that the TX line goes 1 if System loopback is enabled.

pub fn llpbk(&self) -> LLPBK_R

Bit 5 - Line loopback enable. If this bit is turned on, incoming bits are forwarded to TX for testing purpose. See Block Diagram. Note that the internal design sees RX value as 1 always if line loopback is enabled.

pub fn parity_en(&self) -> PARITY_EN_R

Bit 6 - If true, parity is enabled in both RX and TX directions.

pub fn parity_odd(&self) -> PARITY_ODD_R

Bit 7 - If PARITY_EN is true, this determines the type, 1 for odd parity, 0 for even.

pub fn rxblvl(&self) -> RXBLVL_R

Bits 8:9 - Trigger level for RX break detection. Sets the number of character times the line must be low to detect a break.

pub fn nco(&self) -> NCO_R

Bits 16:31 - BAUD clock rate control.