Type Definition earlgrey_registers::uart::ctrl::W

type W = W<u32, CTRL>;

Writer for register CTRL

Methods

impl W

pub fn tx(&mut self) -> TX_W

Bit 0 - TX enable

pub fn rx(&mut self) -> RX_W

Bit 1 - RX enable

pub fn nf(&mut self) -> NF_W

Bit 2 - RX noise filter enable. If the noise filter is enabled, RX line goes through the 3-tap repetition code. It ignores single IP clock period noise.

pub fn slpbk(&mut self) -> SLPBK_W

Bit 4 - System loopback enable. If this bit is turned on, any outgoing bits to TX are received through RX. See Block Diagram. Note that the TX line goes 1 if System loopback is enabled.

pub fn llpbk(&mut self) -> LLPBK_W

Bit 5 - Line loopback enable. If this bit is turned on, incoming bits are forwarded to TX for testing purpose. See Block Diagram. Note that the internal design sees RX value as 1 always if line loopback is enabled.

pub fn parity_en(&mut self) -> PARITY_EN_W

Bit 6 - If true, parity is enabled in both RX and TX directions.

pub fn parity_odd(&mut self) -> PARITY_ODD_W

Bit 7 - If PARITY_EN is true, this determines the type, 1 for odd parity, 0 for even.

pub fn rxblvl(&mut self) -> RXBLVL_W

Bits 8:9 - Trigger level for RX break detection. Sets the number of character times the line must be low to detect a break.

pub fn nco(&mut self) -> NCO_W

Bits 16:31 - BAUD clock rate control.