Type Definition earlgrey_registers::uart::fifo_ctrl::W

type W = W<u32, FIFO_CTRL>;

Writer for register FIFO_CTRL

Methods

impl W

pub fn rxrst(&mut self) -> RXRST_W

Bit 0 - RX fifo reset. Write 1 to the register resets RX_FIFO. Read returns 0

pub fn txrst(&mut self) -> TXRST_W

Bit 1 - TX fifo reset. Write 1 to the register resets TX_FIFO. Read returns 0

pub fn rxilvl(&mut self) -> RXILVL_W

Bits 2:4 - Trigger level for RX interrupts. If the FIFO depth is greater than or equal to the setting, it raises rx_watermark interrupt.

pub fn txilvl(&mut self) -> TXILVL_W

Bits 5:6 - Trigger level for TX interrupts. If the FIFO depth is greater than or equal to the setting, it raises tx_watermark interrupt.